pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "formseal-inbox"
version = "0.1.0"
description = "CLI tool for decrypting formseal ciphertexts"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
authors = [
    {name = "grayguava"}
]
dependencies = [
    "keyring>=25.0",
    "pynacl>=1.5.0",
]

[project.scripts]
fsi = "fsi.fsi:main"

[tool.setuptools]
packages = [
    "fsi",
    "fsi.ui",
    "fsi.commands",
    "fsi.commands.general",
    "fsi.commands.config",
    "fsi.commands.connect",
    "fsi.commands.decrypt",
    "fsi.security",
]